fix-deploy-bugs
