+++ src/Pages/HomePage.jsx
@@ -1,20 +1,22 @@
 import { Header } from 'components/Header/Header';
-import React, { Suspense } from 'react';
-import { Outlet } from 'react-router-dom';
+
 import {  
   // ButtonText, 
   HomeWrapper,
   //  Text 
   } from './HomePage.styled';
+
+import ScreensPage from './ScreensPage';
 // import { useTranslation } from 'react-i18next';
 // import { useSelector } from 'react-redux';
 // import { selectAllBoards } from 'redux/boards/selectors';
 
-import ScreensPage from './ScreensPage';
+// import ScreensPage from './ScreensPage';
 
 export default function HomePage() {
   // const { t } = useTranslation('global');
   // const boards = useSelector(selectAllBoards);
+  // const { boardId } = useParams();
 
   return (
     <HomeWrapper>
@@ -24,9 +26,8 @@ export default function HomePage() {
       <ButtonText>{t('screenPage.static.message2')}</ButtonText>
       {t('screenPage.static.message3')}</Text>} */}
       <ScreensPage/>
-      {/* <Suspense fallback={null}>
-          <Outlet />
-        </Suspense>     */}
-    </HomeWrapper>
+      {/* <Link to={`/home/${boardId}`}></Link> */}
+       {/* <Suspense fallback={<Loader/>}><Outlet /></Suspense>     */}
+    </HomeWrapper> 
   );
 }
