+++ src/components/Button/AddColumnButton.styled.jsx
@@ -59,7 +59,7 @@ export const AddColumnInput = styled.input`
   opacity: 0.4;
   border-radius: ${props => props.theme.radius.m};
   border: 1px solid ${props => props.theme.violetColors.accentColor};
-  box-shadow: 0 4px 16px 0 ${props => props.theme.colors.shadowColor});
+  box-shadow: 0 4px 16px 0 ${props => props.theme.colors.shadowColor};
   background-color: transparent;
   color: ${props => props.theme.colors.darkTextColor};
   letter-spacing: -0.32px;
