+++ src/components/ColumnModal/ColumnModal.css
@@ -1,4 +1,4 @@
-.modal-overlay {
+/* .modal-overlay {
   position: fixed;
   top: 0;
   left: 0;
@@ -8,7 +8,7 @@
   align-items: center;
   flex-direction: column;
   padding: 40px 10px 40px 10px;
-  /* background-color: rgba(0, 0, 0, 0.7) !important; */
+  background-color: rgba(0, 0, 0, 0.7) !important;
   overflow: hidden;
   overflow-y: auto;
   z-index: 999;
@@ -59,4 +59,4 @@
 .ReactModal__Content--before-close {
   transform: translateY(50px);
 }
-
+ */
