+++ src/components/Sidebar/BoardListItem/BoardListItem.styled.js
@@ -1,4 +1,4 @@
-import { Link } from "react-router-dom";
+// import { Link } from "react-router-dom";
 import styled from "styled-components";
 
 export const StyledBoardList = styled.div`
