+++ src/components/boardModals/ModalsBase/BoardModalBase.css
@@ -1,4 +1,4 @@
-.modal-overlay {
+/* .modal-overlay {
     position: fixed;
     top: 0;
     left: 0;
@@ -12,4 +12,4 @@
     overflow-y: auto;
     z-index: 100;
     transition: opacity 0.3s;
-}
+} */
